🚚 Restructure project directories for better modularity
